move `internal` into separate dir
