+++ pkg/server/server_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/MattSScott/basePlatformSOMAS/pkg/internal/testUtils"
+	"github.com/MattSScott/basePlatformSOMAS/internal/testUtils"
 	"github.com/google/uuid"
 )
